Implemented neshta.

+++ Data/Models/Dish.cs
@@ -14,7 +14,7 @@ namespace Data
     {
         public int ID { get; set; }
         public string DishName { get; set; }
-        public double DishPrice { get; set; }
+        public decimal DishPrice { get; set; }
         public string DishDescription { get; set; }
         public byte[] Image { get; set; }
         public bool IsVegetarian { get; set; }
